Use ParamEvaluator class for recursion
