+++ lib/rails_param/param.rb
@@ -1,6 +1,5 @@
-require 'pry'
 module RailsParam
   def param!(name, type, options = {}, &block)
-    MockController.new(params, name, type, options).my_begin(&block)
+    ParamEvaluator.new(params).param!(name, type, options, &block)
   end
 end
